// src/app/models/countdown-model.ts
export enum CountdownDisplayOptions {
  LongRange = 'LongRange',
  MediumRange = 'MediumRange',
  MediumShortRange = 'MediumShortRange',
  ShortRange = 'ShortRange',
  CriticalRange = 'CriticalRange',
}

export enum IntervalElements {
  Years = 'years',
  Months = 'months',
  Weeks = 'weeks',
  Days = 'days',
  Hours = 'hours',
  Minutes = 'minutes',
  Seconds = 'seconds',
}

export enum CountdownCountNumberMilliseconds {
  Second = 1000,
  Minute = 60000,
  Hours = 3600000,
}